ocsipersist-pgsql

For the API documentation see OPAM package ocsipersist.

Here we document the configuration options of the ocsigenserver extension. All attributes of the database tag are optional.

One can either define a host to connect to:

<extension findlib-package="ocsipersist.pgsql">
  <database
     host="localhost"
     port="3000"
     user="ocsipersist"
     password="<secret>"
     size_conn_pool = "16"
  />
</extension>

Or a UNIX domain socket:

<extension findlib-package="ocsipersist.pgsql">
  <database
    unix_domain_socket_dir = "pgsql_socket"
    user="ocsipersist"
    password="<secret>"
    size_conn_pool = "16"
  />
</extension>